[QNN EP] Refactor QLinearMatMul as a node group fusion
